(PC-1281) PR fixes v2
