+++ local_providers/titelive_thing_descriptions.py
@@ -1,12 +1,12 @@
 import re
 from pathlib import PurePath
 
+from connectors.ftp_titelive import get_files_to_process_from_titelive_ftp, get_zip_file_from_ftp
+from domain.titelive import get_date_from_filename, read_description_date
 from models.local_provider import LocalProvider, ProvidableInfo
 from models.local_provider_event import LocalProviderEventType
 from models.thing import Thing
 from repository import local_provider_event_queries
-from utils.ftp_titelive import get_files_to_process_from_titelive_ftp, \
-    get_date_from_filename, read_description_date, get_zip_file_from_ftp
 from utils.logger import logger
 
 DATE_REGEXP = re.compile('Resume(\d{6}).zip')
@@ -31,7 +31,7 @@ class TiteLiveThingDescriptions(LocalProvider):
         self.zips = self.get_remaining_files_to_check(all_zips)
         self.desc_zipinfos = None
         self.zip = None
-        self.dateModified = None
+        self.date_modified = None
 
     def open_next_file(self):
         if self.zip:
@@ -46,7 +46,7 @@ class TiteLiveThingDescriptions(LocalProvider):
 
         self.desc_zipinfos = self.get_description_files_from_zip_info()
 
-        self.dateModified = read_description_date(str(get_date_from_filename(self.zip, DATE_REGEXP)))
+        self.date_modified = read_description_date(str(get_date_from_filename(self.zip, DATE_REGEXP)))
 
     def __next__(self):
         if self.desc_zipinfos is None:
@@ -60,7 +60,7 @@ class TiteLiveThingDescriptions(LocalProvider):
 
         providable_info = ProvidableInfo()
         providable_info.type = Thing
-        providable_info.dateModifiedAtProvider = self.dateModified
+        providable_info.dateModifiedAtProvider = self.date_modified
         path = PurePath(self.desc_zipinfo.filename)
         providable_info.idAtProviders = path.name.split('_', 1)[0]
         self.thingId = providable_info.idAtProviders
