+++ local_providers/titelive_thing_thumbs.py
@@ -2,12 +2,12 @@ import re
 from datetime import datetime
 from pathlib import PurePath
 
+from connectors.ftp_titelive import get_files_to_process_from_titelive_ftp, get_zip_file_from_ftp
+from domain.titelive import get_date_from_filename
 from models.local_provider import LocalProvider, ProvidableInfo
 from models.local_provider_event import LocalProviderEventType
 from models.thing import Thing
 from repository import local_provider_event_queries
-from utils.ftp_titelive import get_files_to_process_from_titelive_ftp, get_date_from_filename, \
-    get_zip_file_from_ftp
 from utils.logger import logger
 
 DATE_REGEXP = re.compile('livres_tl(\d+).zip')
